Moving the plugin initialization into the entrypoint.
