+++ src/file-schema.ts
@@ -1,8 +1,5 @@
 import { z } from "zod";
 
-// @todo should be a better place for this
-import "./metadata"; // required to apply the plugin
-
 export const ezFileBrand = Symbol("File");
 
 const bufferSchema = z.custom<Buffer>((subject) => Buffer.isBuffer(subject), {
