+++ vitest.setup.ts
@@ -1,0 +1,2 @@
+// required to apply the plugin before running the tests, because some tests do not import from entrypoint
+import "./src/zod-plugin";
